Fix i*i overflow of int

+++ src/CountFactors.java
@@ -1,8 +1,8 @@
 class CountFactors {
     public int solution(int N) {
-        
-        int i = 1;
+        long i = 1;
         int result = 0;
+        
         while(i * i < N) {
             if (N % i == 0)
                 result += 2;
